Extract overquota tests
