+++ Makefile
@@ -303,6 +303,7 @@ SPEC_HELPER_MIN_SPECS = \
 	spec/lib/carto/organization_assets_service_spec.rb \
 	spec/lib/carto/storage_options/local_spec.rb \
 	spec/models/carto/username_proposer_spec.rb \
+	spec/services/carto/overquota_users_service_spec.rb \
 	$(NULL)
 
 # This class must be tested isolated as pollutes namespace
