+++ app/services/carto/overquota_users_service.rb
@@ -1,28 +1,35 @@
 module Carto
   class OverquotaUsersService
     def initialize(date)
-      @date_key = "overquota:users:#{date.strftime('%Y%m%d')}"
+      @date = date
     end
 
     def store_overquota_users(delta)
       overquota(delta).each do |user|
-        $users_metadata.hmset(@date_key, user.id, user.data.to_json)
+        $users_metadata.hmset(date_key, user.id, user.data.to_json)
       end
       # Expire the cache after two months. Enough time to review data if needed
-      $users_metadata.expire(@date_key, 2.months)
+      $users_metadata.expire(date_key, 2.months)
     end
 
     def get_stored_overquota_users
-      @date_key = key(date)
-      if $users_metadata.exists(@date_key)
-        $users_metadata.hgetall(@date_key).values.map { |v| JSON.parse(v) }
+      if $users_metadata.exists(date_key)
+        $users_metadata.hgetall(date_key).values.map { |v| JSON.parse(v) }
       else
-        CartoDB::Logger.warning(message: "There is no overquota cached users for date #{formated_date}")
+        CartoDB::Logger.warning(message: "There is no overquota cached users for date #{formatted_date}")
         []
       end
     end
 
     private
+
+    def date_key
+      "overquota:users:#{formatted_date}"
+    end
+
+    def formatted_date
+      @date.strftime('%Y%m%d')
+    end
 
     ##
     # SLOW! Checks redis data (geocodings and isolines) for every user
@@ -30,7 +37,8 @@ module Carto
     #        example: 0.20 will get all users at 80% of their map view limit
     #
     def overquota(delta)
-      ::User.where(enabled: true, organization_id: nil).exclude(account_type: 'FREE').use_cursor.select do |u|
+      overquota_users = []
+      ::User.where(enabled: true, organization_id: nil).exclude(account_type: 'FREE').use_cursor.each do |u|
         limit = u.geocoding_quota.to_i - (u.geocoding_quota.to_i * delta)
         over_geocodings = u.get_geocoding_calls > limit
 
@@ -49,9 +57,11 @@ module Carto
         limit = u.mapzen_routing_quota.to_i - (u.mapzen_routing_quota.to_i * delta)
         over_mapzen_routing = u.get_mapzen_routing_calls > limit
 
-        over_geocodings || over_twitter_imports || over_here_isolines ||
-          over_mapzen_routing || over_obs_snapshot || over_obs_general
+        if over_geocodings || over_twitter_imports || over_here_isolines || over_mapzen_routing || over_obs_snapshot || over_obs_general
+          overquota_users << u
+        end
       end
+      overquota_users
     end
   end
 end
