+++ spec/models/user_spec.rb
@@ -494,96 +494,6 @@ describe User do
     end
   end
 
-  describe '#overquota' do
-    before(:all) do
-      @user.account_type = 'NOT FREE'
-      @user.save
-      @user2.account_type = 'NOT FREE'
-      @user2.save
-    end
-
-    # Filter overquota users to only those created by this spec
-    def overquota(delta = 0)
-      ::User.overquota(delta).select { |u| [@user.id, @user2.id].include?(u.id) }
-    end
-
-    it "should return users near their geocoding quota" do
-      ::User.any_instance.stubs(:get_api_calls).returns([0])
-      ::User.any_instance.stubs(:map_view_quota).returns(120)
-      ::User.any_instance.stubs(:get_geocoding_calls).returns(81)
-      ::User.any_instance.stubs(:geocoding_quota).returns(100)
-      overquota.should be_empty
-      overquota(0.20).map(&:id).should include(@user.id)
-      overquota(0.20).size.should == 2
-      overquota(0.10).should be_empty
-    end
-
-    it "should return users near their here isolines quota" do
-      ::User.any_instance.stubs(:get_api_calls).returns([0])
-      ::User.any_instance.stubs(:map_view_quota).returns(120)
-      ::User.any_instance.stubs(:get_geocoding_calls).returns(0)
-      ::User.any_instance.stubs(:geocoding_quota).returns(100)
-      ::User.any_instance.stubs(:get_here_isolines_calls).returns(81)
-      ::User.any_instance.stubs(:here_isolines_quota).returns(100)
-      overquota.should be_empty
-      overquota(0.20).map(&:id).should include(@user.id)
-      overquota(0.20).size.should == 2
-      overquota(0.10).should be_empty
-    end
-
-    it "should return users near their data observatory snapshot quota" do
-      ::User.any_instance.stubs(:get_api_calls).returns([0])
-      ::User.any_instance.stubs(:map_view_quota).returns(120)
-      ::User.any_instance.stubs(:get_geocoding_calls).returns(0)
-      ::User.any_instance.stubs(:geocoding_quota).returns(100)
-      ::User.any_instance.stubs(:get_here_isolines_calls).returns(0)
-      ::User.any_instance.stubs(:here_isolines_quota).returns(100)
-      ::User.any_instance.stubs(:get_obs_general_calls).returns(0)
-      ::User.any_instance.stubs(:obs_general_quota).returns(100)
-      ::User.any_instance.stubs(:get_obs_snapshot_calls).returns(81)
-      ::User.any_instance.stubs(:obs_snapshot_quota).returns(100)
-      overquota.should be_empty
-      overquota(0.20).map(&:id).should include(@user.id)
-      overquota(0.20).size.should == 2
-      overquota(0.10).should be_empty
-    end
-
-    it "should return users near their data observatory general quota" do
-      ::User.any_instance.stubs(:get_api_calls).returns([0])
-      ::User.any_instance.stubs(:map_view_quota).returns(120)
-      ::User.any_instance.stubs(:get_geocoding_calls).returns(0)
-      ::User.any_instance.stubs(:geocoding_quota).returns(100)
-      ::User.any_instance.stubs(:get_here_isolines_calls).returns(0)
-      ::User.any_instance.stubs(:here_isolines_quota).returns(100)
-      ::User.any_instance.stubs(:get_obs_snapshot_calls).returns(0)
-      ::User.any_instance.stubs(:obs_snapshot_quota).returns(100)
-      ::User.any_instance.stubs(:get_obs_general_calls).returns(81)
-      ::User.any_instance.stubs(:obs_general_quota).returns(100)
-      overquota.should be_empty
-      overquota(0.20).map(&:id).should include(@user.id)
-      overquota(0.20).size.should == 2
-      overquota(0.10).should be_empty
-    end
-    it "should return users near their twitter quota" do
-      ::User.any_instance.stubs(:get_api_calls).returns([0])
-      ::User.any_instance.stubs(:map_view_quota).returns(120)
-      ::User.any_instance.stubs(:get_geocoding_calls).returns(0)
-      ::User.any_instance.stubs(:geocoding_quota).returns(100)
-      ::User.any_instance.stubs(:get_twitter_imports_count).returns(81)
-      ::User.any_instance.stubs(:twitter_datasource_quota).returns(100)
-      overquota.should be_empty
-      overquota(0.20).map(&:id).should include(@user.id)
-      overquota(0.20).size.should == 2
-      overquota(0.10).should be_empty
-    end
-
-    it "should not return organization users" do
-      ::User.any_instance.stubs(:organization_id).returns("organization-id")
-      ::User.any_instance.stubs(:organization).returns(Organization.new)
-      overquota.should be_empty
-    end
-  end
-
   describe '#private_maps_enabled?' do
     it 'should not have private maps enabled by default' do
       user_missing_private_maps = create_user :email => 'user_mpm@example.com',  :username => 'usermpm',  :password => 'usermpm'
